requirements.txt
# PDF Processing
pymupdf

# OCR Engines  
paddlepaddle
paddleocr
easyocr

# Image Processing
opencv-python
pillow
numpy

# Gemini API
google-generativeai

# Utilities
python-dateutil
google-generativeai>=0.3.0
python-dotenv>=1.0.0
groq>=0.4.0

#Dashboard
streamlit==1.30.0
plotly==5.18.0